Break serve/http.rs into separate files.
The primary API handler is currently embedded in commands/serve/http.rs
and is quite big. This extracts the shared types from the http.rs into
its own types.rs file (similar to what we did with the store module),
and makes http/ a module directory.

We'll be taking this even further by moving the http layer entirely.
