+++ src/commands/serve/http/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (c) 2025 Chris Corbyn <chris@zizq.io>
+// Licensed under the Business Source License 1.1. See LICENSE file for details.
+
+mod http;
+mod types;
+
+pub use http::app;
